compile the sql strings into code at compile time.
This makes them available at run time for both a jar and a dev repl.
The downside is you have to recompile the macro if you update the file,
but there are workarounds.

this is wip still as i havent tested build-all and i think some reporting
functionality got lost.